backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.3.0
numpy>=1.24.0
